Linux 2.2.19pre14
o	Update slhc code for endianness			(Dave Miller)
o	Update s390 dasd driver				(Ulrich Weigand)
o	Allow more than 4K of partitions		(Ulrich Weigand)
o	Fix check in sockfilter				(Dave Miller)
o	Sparc updates (quad sbus sunhme detect, BUG())	(Dave Miller)
o	Fix hid locking and ston32 bugs			(Paul Mackerras)
o	Update 3c59x drivers		(Andrew Morton, Maciej Rozycki,
					 Fred Maciel, Georg Engstrand,
					 Brett Frankenberger, Don Becker,
o	Fix a usb message				(Randy Dunlap)
o	Eicon driver updates				(Armin Schindler)
o	Update 8139too driver				(Jens David)
o	Fix USB hub locks				(Paul Mackerras)
o	Fix missing keyspan config line			(Paul Mackerras)
o	Merge S/390 bug fixes				(Ulrich Weigand)
o	Some S/390 cleanups				(Ulrich Weigand)
o	Update S/390 ELF magic 				(Ulrich Weigand)
o	Update hwc driver				(Ulrich Weigand)
o	Update ctc driver				(Ulrich Weigand)
o	Update iucv driver				(Ulrich Weigand)
o	S/390 warning fixes				(Ulrich Weigand)

+++ net/core/filter.c
@@ -196,7 +196,7 @@ int sk_run_filter(struct sk_buff *skb, struct sock_filter *filter, int flen)
 			case BPF_LD|BPF_W|BPF_ABS:
 				k = fentry->k;
 load_w:
-				if(k+sizeof(u32) <= len) {
+				if((unsigned int)(k+sizeof(u32)) <= len) {
 					A = ntohl(*(u32*)&data[k]);
 					continue;
 				}
@@ -215,7 +215,7 @@ load_w:
 			case BPF_LD|BPF_H|BPF_ABS:
 				k = fentry->k;
 load_h:
-				if(k + sizeof(u16) <= len) {
+				if((unsigned int) (k + sizeof(u16)) <= len) {
 					A = ntohs(*(u16*)&data[k]);
 					continue;
 				}
@@ -234,7 +234,7 @@ load_h:
 			case BPF_LD|BPF_B|BPF_ABS:
 				k = fentry->k;
 load_b:
-				if(k < len) {
+				if((unsigned int)k < len) {
 					A = data[k];
 					continue;
 				}
@@ -271,7 +271,7 @@ load_b:
 
 			case BPF_LDX|BPF_B|BPF_MSH:
 				k = fentry->k;
-				if(k >= len)
+				if((unsigned int)k >= len)
 					return (0);
 				X = (data[k] & 0xf) << 2;
 				continue;
